Moved main view logic
